BSGenome.Dsimulans.NCBI.3.1.fa {BSGenome.Dsimulans.NCBI.3.1.fa}	R Documentation

Full genome sequences for Drosophila simulans (NCBI v. 3.1)

Description

Drosophila simulans full genome sequences (V3.1, Oct. 2021)

Note

This BSgenome data package was made from the following source data files:

GCF_016746395.2.sorted.2bit
  

See ?BSgenomeForge and the BSgenomeForge vignette (vignette("BSgenomeForge")) in the BSgenome software package for how to make a BSgenome data package.

Author(s)

The Bioconductor Dev Team

See Also

BSgenome objects and the available.genomes function in the BSgenome software package.

DNAString objects in the Biostrings package.

The BSgenomeForge vignette (vignette("BSgenomeForge")) in the BSgenome software package for how to make a BSgenome data package.

Examples

BSGenome.Dsimulans.NCBI.3.1.fa
genome <- BSGenome.Dsimulans.NCBI.3.1.fa
head(seqlengths(genome))

## ---------------------------------------------------------------------
## Genome-wide motif searching
## ---------------------------------------------------------------------
## See the GenomeSearching vignette in the BSgenome software
## package for some examples of genome-wide motif searching using
## Biostrings and the BSgenome data packages:
if (interactive())
    vignette("GenomeSearching", package="BSgenome")

[Package BSGenome.Dsimulans.NCBI.3.1.fa version 3.1 Index]